handle explicit remote unmute, removed mute status match
