+++ src/index.ts
@@ -1,4 +1,5 @@
-import { DataPacket_Kind, VideoQuality } from './proto/livekit_rtc';
+import { DataPacket_Kind } from './proto/livekit_models';
+import { VideoQuality } from './proto/livekit_rtc';
 import LocalParticipant from './room/participant/LocalParticipant';
 import Participant from './room/participant/Participant';
 import RemoteParticipant from './room/participant/RemoteParticipant';
